+++ crates/objc2/CHANGELOG.md
@@ -50,6 +50,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `MethodImplementation` to `Return` and `Arguments`.
 * **BREAKING**: Make `rc::Allocated` allowed to be `NULL` internally, such
   that uses of `Option<Allocated<T>>` is now simply `Allocated<T>`.
+* `AnyObject::class` now returns a `'static` reference to the class.
 
 ### Deprecated
 * Soft deprecated using `msg_send!` without a comma between arguments (i.e.